Added Documentation and updated file header

+++ HSL_main.cpp
@@ -1,3 +1,20 @@
+/*
+ * This file is part of the HSL distribution (https://github.com/kristian80/HSL).
+ * Copyright (c) 2019 Kristian80.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, version 3.
+ *
+ * This program is distributed in the hope that it will be useful, but
+ * WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+ * General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
+
 #include <stdio.h>
 #include <string.h>
 #include <stdlib.h>
@@ -301,7 +318,7 @@ int WrapReadStringCallback(
 		strcpy((char*)outValue, pStr->c_str());
 		return 2048;
 	}
-	return pStr->length();
+	return (int) pStr->length();
 
 }
